% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addfiglab.R
\name{addfiglab}
\alias{addfiglab}
\title{Add a figure label in topleft-most corner}
\usage{
addfiglab(lab, font = 2, units = "ndc", cex = 1, ...)
}
\arguments{
\item{lab}{Character giving the figure label}

\item{font}{the color and (if \code{vfont = NULL}) font to be
    used, possibly vectors.  These default to the values of the global
    \link[graphics]{graphical parameters} in \code{\link[graphics]{par}()}.}

\item{units}{Character giving the device units to use -- will typically be
'ndc' or 'nfc'}

\item{cex}{numeric \bold{c}haracter \bold{ex}pansion factor; multiplied
    by \code{\link[graphics]{par}("cex")} yields the final character size.
    \code{NULL} and \code{NA} are equivalent to \code{1.0}.}

\item{...}{Additional parameters passed to text}
}
\description{
Add a figure label in topleft-most corner
}
\details{
Note, 'xpd' is \code{NA} and 'adj' is {c(0, 1)}.
}
