% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addfiglab_line.R
\name{addfiglab_line}
\alias{addfiglab_line}
\title{Add a figure label}
\usage{
addfiglab_line(lab, xl = NULL, yl = NULL, outer = FALSE, ...)
}
\arguments{
\item{lab}{Character giving the figure label}

\item{xl}{Number of lines into the left x margin}

\item{outer}{a logical value indicating whether line refers to the outer 
plot margin, rather than the standard plot margin}

\item{...}{Additional parameters passed to text}

\item{xy}{Number of lines intot the top y margin}
}
\description{
Add a figure label based on x and y line values
}
\details{
Note, 'xpd' is \code{NA} and 'adj' is {c(0, 1)}.
}
