% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha2opaque.R
\name{alpha2opaque}
\alias{alpha2opaque}
\title{Convert transparent color to opaque equivalent}
\usage{
alpha2opaque(c1, c2 = "#FFFFFF")
}
\arguments{
\item{c1}{transparent color, as defined by any of the three R specifications}

\item{c2}{opaque background color, defaults to "white"}
}
\description{
Convert transparent color to opaque equivalent
}
\seealso{
\code{\link[grDevices]{col2rgb}}, \code{\link[grDevices]{rgb}}

\code{\link{col2alpha}}
}
