% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMCC.R
\name{calcMCC}
\alias{calcMCC}
\title{Calculate the Matthew's correlation coefficient}
\usage{
calcMCC(tp, tn, fp, fn)
}
\arguments{
\item{tp}{Number of true positives}

\item{tn}{Number of true negatives}

\item{fp}{Number of false positives}

\item{fn}{Number of false negatives}
}
\description{
Calculate the Matthew's correlation coefficient
}
\references{
\url{https://en.wikipedia.org/wiki/Matthews_correlation_coefficient}
}
