% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFuncs.R
\name{checkFuncs}
\alias{checkFuncs}
\alias{l1chr}
\alias{l1num}
\alias{l1int}
\alias{l1log}
\alias{allSameLength}
\title{Common parameter checks}
\usage{
l1chr(x)

l1num(x)

l1int(x)

l1log(x)

allSameLength(...)
}
\arguments{
\item{x}{R object}
}
\description{
Common function parameter checks
}
\section{Functions}{
\itemize{
\item \code{l1chr}: length 1 character

\item \code{l1num}: length 1 numeric

\item \code{l1int}: length 1 integer

\item \code{l1log}: length 1 logical

\item \code{allSameLength}: all same length
}}

