% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2alpha.R
\name{col2alpha}
\alias{col2alpha}
\title{Modify the alpha value for given color vector}
\usage{
col2alpha(col, alpha = 0.5)
}
\arguments{
\item{col}{vector of colors as defined by any of the three R specifications}

\item{alpha}{numeric of length 1, between 0 and 1}
}
\description{
Modify the alpha value for given color vector
}
\seealso{
\code{\link[grDevices]{col2rgb}}, \code{\link[grDevices]{rgb}}

\code{\link{alpha2opaque}}
}
