% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinePDF.R
\name{combinePDF}
\alias{combinePDF}
\title{Combine list of pdf files}
\usage{
combinePDF(input, output)
}
\arguments{
\item{input}{Vector with paths to pdf files}

\item{output}{Character with path to output pdf}
}
\description{
Combine list of pdf files
}
\details{
This function utilizes ghostscript, and will not work without a Ghostscript
installation
}
\note{
The strength of this approach is preserving hyperlinks.
}
