% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCall.R
\name{doCall}
\alias{doCall}
\title{Wrapper for do.call}
\usage{
doCall(f, args, ..., quote = FALSE, envir = parent.frame(), warnDup = TRUE)
}
\arguments{
\item{args}{a \emph{list} of arguments to the function call.  The
    \code{names} attribute of \code{args} gives the argument names.}

\item{...}{Additional named arguments to be included in 'args'}

\item{quote}{a logical value indicating whether to quote the
    arguments.}

\item{envir}{an environment within which to evaluate the call.  This
    will be most useful if \code{what} is a character string and
    the arguments are symbols or quoted expressions.}

\item{warnDup}{logical}
}
\value{
The result of the (evaluated) function call.
}
\description{
Wrapper for 'do.call' with the shorthand for additional args
}
\details{
When arguments are duplicated between 'args' and '...' preference is given
to '...' (with a warning when warnDup = TRUE, the default).
}
\examples{
demolist <- list(a = 1:10, b = "hello")
demofunc <- function(a, b, c) list(a, b, c)
## The following errors
## do.call(demofunc, c(demolist, c = 1:2)) 
do.call(demofunc, c(demolist, list(c = 1:2)))
doCall(demofunc, demolist, c = 1:2)
}
