% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitchPlots.R
\name{fig2minipage}
\alias{fig2minipage}
\title{Modify the alpha value for given color vector}
\usage{
fig2minipage(path, width, newline = FALSE, adjWidth = FALSE, ...)
}
\arguments{
\item{path}{file path to pdf for include graphics}

\item{width}{numeric of length 1, between 0 and 1 -- passed to minipage width}

\item{newline}{logical of length 1, when FALSE a '\%' is appended to the end
minipage call}

\item{adjWidth}{logical of length 1, when TRUE [width=\\linewidth] is added
to the includegraphics call}

\item{...}{added for flexibility, not passed to anything}
}
\description{
Modify the alpha value for given color vector
}
\note{
NOT exported
}
\seealso{
\code{\link{stitchPlots}}
}
