% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthFuncs.R
\name{lengthFuncs}
\alias{lengthFuncs}
\alias{lu}
\alias{lw}
\title{Common length calculations}
\usage{
lu(x)

lw(x)
}
\arguments{
\item{x}{R object}
}
\description{
Common length calculations
}
\section{Functions}{
\itemize{
\item \code{lu}: length(unique(x))

\item \code{lw}: sum(x, na.rm = TRUE)
}}

