% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line2user.R
\name{line2user}
\alias{line2user}
\title{Get margin line location in usr coordinates}
\usage{
line2user(line, side, outer = FALSE)
}
\arguments{
\item{line}{the number of lines into the margin}

\item{side}{an integer specifying which side of the plot: 
1=below, 2=left, 3=above and 4=right.}

\item{outer}{a logical value indicating whether line refers to the outer 
plot margin, rather than the standard plot margin}
}
\description{
Get margin line location in usr coordinates
}
\details{
Code credit: \url{https://stackoverflow.com/a/30835971/1623354}; modification
for 'outer' parameter added by DLF.
}
