% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2dDen.R
\name{log2dDen}
\alias{log2dDen}
\title{Calculate 2D densities on a log scale}
\usage{
log2dDen(x, y, nbin = 100, lims = c(range(x), range(y)), freq = FALSE)
}
\arguments{
\item{x}{Numeric input for x dimension}

\item{y}{Numeric input for y dimension}

\item{nbin}{Integer of length 1, the number of bins in each dimension}

\item{lims}{Numeric of length 4, ranges for the x and y values, respectively}

\item{freq}{Logical of length 1, counts converted to proportions when FALSE}
}
\value{
list with x, y: seq(1, nbin) for plotting; z: nbin x nbin matrix 
with the freq/proportion values at the x and y bins; xseq, yseq: numeric
vectors containing the cut points for the x and y bins.
}
\description{
Calculate 2D densities on a log scale
}
