% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logSumExp.R
\name{logSumExp}
\alias{logSumExp}
\title{Get posterior probabilities for a vector of log-likelihoods}
\usage{
logSumExp(l)
}
\arguments{
\item{l}{Numeric vector of log-likelihoods}
}
\description{
Get posterior probabilities for a vector of log-likelihoods
}
