% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPal.R
\name{plotPal}
\alias{plotPal}
\title{Plot palette of colors}
\usage{
plotPal(col = palette())
}
\arguments{
\item{col}{The colors to plot; plots current value of palette() if NULL}
}
\description{
Plot palette of colors
}
