% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantHist.R
\name{quantHist}
\alias{quantHist}
\title{Draw histogram of reduced range, by lower and upper quantile}
\usage{
quantHist(x, lq = 0.05, uq = 0.95, ...)
}
\arguments{
\item{x}{Numeric input to hist}

\item{lq}{Lower quantile used to subset x}

\item{uq}{Upper quantile used to subset x}
}
\description{
Draw histogram of reduced range, by lower and upper quantile
}
