% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce3mad.R
\name{reduce3mad}
\alias{reduce3mad}
\title{Reduce given data to med +/- 3mad}
\usage{
reduce3mad(x, returnBound = FALSE)
}
\arguments{
\item{x}{Numeric}

\item{returnBound}{Logical of length 1, return bounds rather than reduced set}
}
\description{
Reduce given data to med +/- 3mad
}
