% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmArray.R
\name{slurmArray}
\alias{slurmArray}
\title{Wrapper for rslurm::slurm_apply}
\usage{
slurmArray(f, params, jobname = NA, mem = 10000L, maxJobs = 500L,
  cpu = 1L, add_objects = NULL, pkgs = rev(.packages()),
  submit = TRUE, libPaths = NULL)
}
\arguments{
\item{f}{A function that accepts one or many single values as parameters and
may return any type of R object.}

\item{params}{A data frame of parameter values to apply \code{f} to. Each
column corresponds to a parameter of \code{f} (\emph{Note}: names must
match) and each row corresponds to a separate function call.}

\item{jobname}{The name of the Slurm job; if \code{NA}, it is assigned a
random name of the form "slr####".}

\item{mem}{integer of length 1, how much memory to allocate to each job in
bytes}

\item{maxJobs}{integer of length 1, what is the maximum jobs to run at one 
time?}

\item{cpu}{integer of length 1, how many cores for each job?}

\item{add_objects}{A character vector containing the name of R objects to be
saved in a .RData file and loaded on each cluster node prior to calling
\code{f}.}

\item{pkgs}{A character vector containing the names of packages that must
be loaded on each cluster node. By default, it includes all packages
loaded by the user when \code{slurm_apply} is called.}

\item{submit}{Whether or not to submit the job to the cluster with
\code{sbatch}; see Details below for more information.}

\item{libPaths}{A character vector describing the location of additional R
library trees to search through, or NULL. The default value of NULL
corresponds to libraries returned by \code{.libPaths()} on a cluster node.
Non-existent library trees are silently ignored.}
}
\value{
A \code{slurm_job} object containing the \code{jobname} and the
  number of \code{nodes} effectively used.
}
\description{
Shorthand wrapper for submitting slurm job arrays using 
\code{\link[rslurm]{slurm_apply}}
}
\details{
Submits a slurm array with a job for each row in params. See 
\code{\link[rslurm]{slurm_apply}} for more details.
}
\seealso{
\code{\link[rslurm]{slurm_apply}}
}
