% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitchPlots.R
\name{stitchPlots}
\alias{stitchPlots}
\title{Combine PDFs into a single plot using minipages}
\usage{
stitchPlots(figLst, width, height, figname = "Rplot.pdf")
}
\arguments{
\item{figLst}{list of "figures" -- see details}

\item{width}{numeric of length 1, width of the output pdf in inches}

\item{height}{numeric of length 1, height of the output pdf in inches}

\item{figname}{character of length 1, output file name/path}
}
\description{
Combine PDFs into a single plot using minipages
}
\details{
Each item in 'figLst' points to the input PDF and provides details about the
minipage environment. Each item should include
\itemize{
\item[path] character of length 1, the path to the PDF/image, passed to 
includegraphics
\item[width] numeric of length 1, the width passed to minipage as a fraction 
of linewidth (0-1]
\item[newline] logical of length 1, should the minipage be followed by a 
newline or escaped using a '\%'; newline when TRUE, '%' when FALSE
\item[adjWidth] logical of length 1, should the images be adjusted to the
linewidth of the bounding minipage
}

Figures are all centered, and included in the list order. PDF compiled by
\code{\link[tinytex]{xelatex}} in a temporary directory.
}
\seealso{
\code{\link[tinytex]{xelatex}}, \code{\link{fig2minipage}}
}
