% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addfiglab.R
\name{addfiglab}
\alias{addfiglab}
\title{Add a figure label}
\usage{
addfiglab(lab, xl = par()$mar[2], yl = par()$mar[3], cex = 1, ...)
}
\arguments{
\item{lab}{Character giving the figure label}

\item{xl}{Number of lines into the left x margin}

\item{cex}{A numerical value giving the amount by which figure label 
should be magnified relative to the default.}

\item{...}{Additional parameters passed to text}

\item{xy}{Number of lines intot the top y margin}
}
\description{
Add a figure label based on x and y line values
}
\details{
Note, 'xpd' is \code{NA} and 'adj' is {c(0, 1)}.
}
