% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addfiglab.R
\name{addfiglab}
\alias{addfiglab}
\title{Add a figure label in topleft-most corner}
\usage{
addfiglab(lab, font = 2, units = "ndc", ...)
}
\arguments{
\item{lab}{Character giving the figure label}

\item{font}{the color and (if \code{vfont = NULL}) font to be
    used, possibly vectors.  These default to the values of the global
    \link{graphical parameters} in \code{\link{par}()}.}

\item{units}{Character giving the device units to use -- will typically be
'ndc' or 'nfc'}

\item{...}{Additional parameters passed to text}
}
\description{
Add a figure label in topleft-most corner
}
\details{
Note, 'xpd' is \code{NA} and 'adj' is {c(0, 1)}.
}
